Move nested helper TryGetNavigateToSearchService to the containing class
