+++ src/EditorFeatures/Core.Wpf/NavigateTo/NavigateToItemProvider.cs
@@ -4,6 +4,8 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.CodeAnalysis.NavigateTo;
 using Microsoft.CodeAnalysis.Shared.TestHooks;
 using Microsoft.VisualStudio.Language.NavigateTo.Interfaces;
 using Roslyn.Utilities;
@@ -158,6 +160,48 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.NavigateTo
                 _cancellationTokenSource.Token);
 
             searcher.Search();
+        }
+
+        private static INavigateToSearchService_RemoveInterfaceAboveAndRenameThisAfterInternalsVisibleToUsersUpdate TryGetNavigateToSearchService(Project project)
+        {
+            var service = project.LanguageServices.GetService<INavigateToSearchService_RemoveInterfaceAboveAndRenameThisAfterInternalsVisibleToUsersUpdate>();
+            if (service != null)
+            {
+                return service;
+            }
+
+#pragma warning disable CS0618 // Type or member is obsolete
+#pragma warning disable CS0612 // Type or member is obsolete
+            var legacyService = project.LanguageServices.GetService<INavigateToSearchService>();
+            if (legacyService != null)
+            {
+                return new ShimNavigateToSearchService(legacyService);
+            }
+#pragma warning restore CS0612 // Type or member is obsolete
+#pragma warning restore CS0618 // Type or member is obsolete
+
+            return null;
+        }
+
+        [Obsolete("https://github.com/dotnet/roslyn/issues/28343")]
+        private class ShimNavigateToSearchService : INavigateToSearchService_RemoveInterfaceAboveAndRenameThisAfterInternalsVisibleToUsersUpdate
+        {
+            private readonly INavigateToSearchService _navigateToSearchService;
+
+            public ShimNavigateToSearchService(INavigateToSearchService navigateToSearchService)
+            {
+                _navigateToSearchService = navigateToSearchService;
+            }
+
+            public IImmutableSet<string> KindsProvided => ImmutableHashSet.Create<string>(StringComparer.Ordinal);
+
+            public bool CanFilter => false;
+
+            public Task<ImmutableArray<INavigateToSearchResult>> SearchDocumentAsync(Document document, string searchPattern, IImmutableSet<string> kinds, CancellationToken cancellationToken)
+                => _navigateToSearchService.SearchDocumentAsync(document, searchPattern, cancellationToken);
+
+            public Task<ImmutableArray<INavigateToSearchResult>> SearchProjectAsync(Project project, string searchPattern, IImmutableSet<string> kinds, CancellationToken cancellationToken)
+                => _navigateToSearchService.SearchProjectAsync(project, searchPattern, cancellationToken);
         }
     }
 }
